backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Projeto Django backend"
authors = ["ChristianSAlves <christiansousaalves17@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.0.3"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.1"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
